consolidate event listeners
